Created API for Schedule    * Index route for get all events.    * Destroy route for delete an event by given id.    * Store route for adding a new event.    * Update route for updating an event by given id.

+++ app/Http/Controllers/ScheduleController.php
@@ -25,11 +25,8 @@ class ScheduleController extends Controller
     public function index(Request $request)
     {
 
-        if ($request) {
-            //dd($request);
-        }
 
-        if ($request->ajax()) {
+        if ($request->ajax() || $request->wantsJson()) {
             $events = Event::latest()->get();
 
             //dd($data);
@@ -58,7 +55,7 @@ class ScheduleController extends Controller
     public function store(Request $request)
     {
 
-        //dd($request->all());
+
         $validator = Validator::make($request->all(), [
             'title' => 'required',
             'start' => 'required',
@@ -73,21 +70,29 @@ class ScheduleController extends Controller
 
 
         //dd($request);
-        $request = $request->all();
-        $request["assigned_by"] = Auth::id();
-        $request["description"] = "No Descripition";
+        $requestArray = $request->all();
+        $requestArray["assigned_by"] = Auth::id();
+        $requestArray["description"] = "No Descripition";
 
 
-        
-        if (empty($request["id"])) {
+
+        if (empty($requestArray["id"])) {
             //dd($request);
-            Event::create($request);
+            Event::create($requestArray);
+            if($request->wantsJson()){
+                return response()->json(["Message" => "Created successfully"]);
+            }
         } else {
             //dd($request);
             // Event::where('id',$request["id"])->update($request);
-            $event = Event::findOrFail($request['id']);
-            $event->update($request);
+            $event = Event::findOrFail($requestArray['id']);
+            $event->update($requestArray);
+            if($request->wantsJson()){
+                return response()->json(["Message" => "Updated successfully"]);
+            }
         }
+
+
         return redirect()->back();
     }
 
@@ -99,7 +104,7 @@ class ScheduleController extends Controller
      */
     public function show($id)
     {
-        return "Hello World";
+        //return "Hello World";
     }
 
     /**
@@ -122,7 +127,11 @@ class ScheduleController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        if($request->wantsJson()){
+            $event = Event::findOrFail($request['id']);
+            $event->update($request->all());
+            return response()->json($event);
+        }
     }
 
     /**
@@ -131,24 +140,27 @@ class ScheduleController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Event $event)
+    public function destroy(Request $request)
     {
-        dd($event);
-        $event->delete();
-
-        redirect()->back();
-    }
-
-    public function allEvents()
-    {
-        $events = Event::latest()->get();
-        return response()->json($events, 200);
-    }
-
-    public function deleteEvent($id){
+        if ($request->wantsJson()) {
+            $event = Event::find($request->id);
+            if(!empty($event)){
+                $event->delete();
+                return response()->json($event);    
+            }
+            return response()->json(["Message" => "Event not found"]);
+        }
         
+    }
+
+
+
+    public function deleteEvent($id)
+    {
+
         //dd(Event::findOrFail($id));
         Event::findOrFail($id)->delete();
+
         return redirect()->back();
     }
 }
